delete interface
